// packages/birth-correction/src/types.ts
export interface SignPhysicalProfile {
  sign: string;
  build: string;
  height: string;
  complexion: string;
  faceShape: string;
  eyeCharacter: string;
  hairTendency: string;
  distinguishingFeature: string;
  voiceQuality: string;
  movementStyle: string;
}